+++ src/low-level/bot-info.ts
@@ -20,4 +20,5 @@ export const genericBotInfo: UserFromGetMe = {
   can_manage_bots: false,
   has_topics_enabled: false,
   allows_users_to_create_topics: false,
+  supports_join_request_queries: false,
 };